requirements.txt
python-dotenv>=1.0.1
PyYAML>=6.0.1
requests>=2.32.3
openai>=1.52.0
google-api-python-client>=2.147.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
protobuf>=4.25.8,<6
